requirements.txt
python-dotenv==1.0.0
langchain==0.3.0
langchain-openai==0.2.0
boto3==1.35.0
click==8.1.7
colorama==0.4.6
pydantic==2.9.0
